Refactor code editor context and global provider components
duniya hila di part -1
