+++ client/src/components/CodeEnvironment.jsx
@@ -1,24 +1,21 @@
 import React, { useEffect, useContext, useState } from "react";
-import {
-  Outlet,
-  redirect,
-  useParams,
-  useNavigate,
-  useLocation,
-} from "react-router-dom";
+import { useParams } from "react-router-dom";
 import { Box, Stack, Button, Chip } from "@mui/material";
 import CodeEditor from "./CodeEditor";
 import CodeEditorTop from "./sub-components/CodeEditorTop";
-import Header from "./sub-components/Header";
-import Drawer from "./sub-components/Drawer";
-import FriendCodeEditor from "./FriendCodeEditor";
+
+// // 
 import Output from "./Output";
-import CreateRoom from "./CreateRoom";
+import CreateRoomModal from "./CreateRoomModal";
 import CodeEditorContext from "../context/CodeEditorContext";
 import problems from "../Data/problems";
 import GlobalContext from "../context/GlobalContext";
 import LoginToContinue from "./sub-components/LoginToContinue";
 import LoginForm from "./LoginForm";
+import QuestionDataDisplay from "./QuestionDisplay/QuestionDataDisplay";
+import CreateRoomButton from "./create room/CreateRoomButton";
+import FriendCodeDrawer from "./create room/FriendCodeDrawer";
+
 const { io } = require("socket.io-client");
 
 const CodeEnviornment = () => {
@@ -26,19 +23,16 @@ const CodeEnviornment = () => {
   const [isOpen, setIsOpen] = useState(false);
   const [displayName, setDisplayName] = useState("");
   const [roomId, setRoomId] = useState("");
-  const [showModal, setShowModal] = useState(false);
+  const { showCreateRoomModal, setShowCreateRoomModal, setFriendText } =
+    useContext(GlobalContext);
   const [isCredentialsEntered, setIsCredentialsEntered] = useState(false);
   const [socket, setSocket] = useState(null);
-  const [friendText, setFriendText] = useState("");
+
   const [friendLanguage, setFriendLanguage] = useState("cpp");
   const { roomCreated, setRoomCreated } = useContext(CodeEditorContext);
 
-  const { loggedIn, setLoggedIn } = useContext(GlobalContext);
-  const [showModal2, setShowModal2] = useState(false);
-  const [showLoginForm, setShowLoginForm] = useState(false);
-  const navigate = useNavigate();
-  // const location = useLocation();
-  // console.log(location.pathname);
+  // const [showLoginForm, setShowLoginForm] = useState(false);
+
   let problemIndex = 0;
   for (let i = 0; i < problems.length; i++) {
     if (problems[i].id === parseInt(problemId)) {
@@ -60,16 +54,8 @@ const CodeEnviornment = () => {
     // Emit 'joinRoom' event to the server with display name and room ID
     newSocket.emit("joinRoom", { displayName, roomId });
     console.log("Connected to room", roomId);
-    setShowModal(false);
+    setShowCreateRoomModal(false);
     setRoomCreated(true);
-  };
-
-  const handleCreateRoom = () => {
-    if (loggedIn) {
-      setShowModal(true);
-    } else {
-      setShowModal2(true);
-    }
   };
 
   useEffect(() => {
@@ -93,8 +79,6 @@ const CodeEnviornment = () => {
 
   useEffect(() => {
     (function () {
-      "use strict";
-
       // horizontal direction
       (function resizableX() {
         const resizer = document.querySelector(".resizer-x");
@@ -245,50 +229,16 @@ const CodeEnviornment = () => {
 
   return (
     <>
-      {showLoginForm && <LoginForm />}
+      {/* {showLoginForm && <LoginForm />} */}
       {/* <LoginForm isOpen={showLoginForm} /> */}
-      <Header isOpen={isOpen} setIsOpen={setIsOpen} />
       <div className="absolute right-0 p-4">
         {" "}
-        {/* Use Tailwind CSS classes to style the container */}
-        {!socket && (
-          <Button
-            //white text, blue background, rounded corners, padding 2px, border 1px solid blue
-            sx={{
-              color: "white",
-              background: "#1976d2",
-              "&:hover": {
-                background: "#225281",
-              },
-              borderRadius: "8px",
-              p: 1,
-              border: "1px solid #1e1e1e",
-            }}
-            onClick={() => handleCreateRoom()}
-            variant="outlined"
-          >
-            Create Room
-          </Button>
 
-          // same button with tailwind classes
-          // <button
-          //   className="text-white bg-blue-500 rounded px-2 border border-blue-500 "
-          //   onClick={() => handleCreateRoom()}
-          // />
-        )}
-        <LoginToContinue
-          isOpen={showModal2}
-          onClose={() => setShowModal2(false)}
-          onLogin={() => {
-            navigate("/login", { replace: true });
-            // setShowModal(true);
-
-            // {showLoginForm && <LoginForm />}
-          }}
-        />
-        <CreateRoom
-          isOpen={showModal}
-          onClose={() => setShowModal(false)}
+        <CreateRoomButton />
+        
+        <CreateRoomModal
+          isOpen={showCreateRoomModal}
+          onClose={() => setShowCreateRoomModal(false)}
           onCreateRoom={connectToRoom}
           displayName={displayName}
           setDisplayName={setDisplayName}
@@ -308,80 +258,16 @@ const CodeEnviornment = () => {
         padding={{ xs: 1, sm: 2, md: 2 }}
         className="resizable-x"
       >
-        <div id="app">
+        <div id="app" classname="h-screen">
           <div className="resizable-x">
             <div style={{ flex: "40%" }}>
-              <div className="quest whitespace-pre-line text-justify h-screen overflow-y-scroll">
-                <div className="flex items-center ">
-                  <Chip
-                    className={`text-sm font-bold px-6 py-4 whitespace-nowrap`}
-                    color={
-                      problem.difficulty === "Easy"
-                        ? "success"
-                        : problem.difficulty === "Medium"
-                        ? "warning"
-                        : "error"
-                    }
-                    label={problem["difficulty"]}
-                    variant="filled"
-                    sx={{ color: "white" }}
-                  />
-                  <p className="p ml-5">
-                    {problem["tags"].map((tag, index) => (
-                      <Chip
-                        key={index}
-                        className="bg-gray m-1"
-                        variant="outlined"
-                        label={tag}
-                        sx={{ color: "white" }}
-                      />
-                    ))}
-                  </p>
-                </div>
-                <br />
-                <div className="pr-5">
-                  <p>{problem["description"]}</p>
-                  <br />
-                  {/* h6 not working */}
-                  <h6 className="text-10xl">
-                    <b>Input</b>
-                  </h6>
-                  <p>{problem["Input"]}</p>
-                  <br />
-                  <h6>
-                    <b>Output</b>
-                  </h6>
-                  <p>{problem["Output"]}</p>
-                  <br />
-                  <div>
-                    <h6>
-                      <b>Example</b>
-                    </h6>
-                    <b>Input</b>
-                    <p>{problem["RunInput"]}</p>
-                    <br />
-                    <b>Output</b>
-                    <p>{problem["RunOutput"]}</p>
-                    <br />
-                  </div>
-                  <h6>
-                    <b>Explanation</b>
-                  </h6>
-                  <p>{problem["Explanation"]}</p>
-                </div>
-              </div>
-              {/* </Box> */}
+              <QuestionDataDisplay />
+          
             </div>
             <div className="resizer-x" />
             <div className="resizable-x" style={{ flex: "60%" }}>
               <div className="div1" style={{ flex: "100%" }}>
-                {/* <p>div 1</p> */}
-                {/* <Box
-                  sx={{
-                    width: "100%",
-                  }}
-                  className="div1"
-                > */}
+         
                 <CodeEditorTop
                   socket={socket}
                   roomId={roomId}
@@ -394,18 +280,15 @@ const CodeEnviornment = () => {
                 />
                 <div className="resizer-y"></div>
                 <div className="resizable-y" style={{ flex: "60%" }}>
-                  <Output />
+               
                 </div>
 
-                {/* </Box> */}
               </div>
             </div>
           </div>
         </div>
       </Stack>
-      <Drawer isOpen={isOpen} setIsOpen={setIsOpen} language={friendLanguage}>
-        <FriendCodeEditor friendText={friendText} language={friendLanguage} />
-      </Drawer>
+      <FriendCodeDrawer />
     </>
   );
 };
